requirements.txt
numpy
tqdm
scipy==1.5.3
torch==1.13.1
torchvision==0.14.1
scikit-image==0.19.2
scikit-lean==1.0.2
pillow==9.4.0
glob2=0.7


